@@ -7,6 +7,8 @@
   "author": "Christopher Best <bestchris@gmail.com>",
   "license": "MIT",
   "scripts": {
+    "lint": "eslint ./src",
+    "lint:fix": "eslint --fix ./src && yarn -s prettier --write ./src",
     "start": "webpack-dev-server --config webpack.dev.js",
     "build:dev": "webpack --config webpack.dev.js",
     "build": "webpack --config webpack.prod.js",
@@ -16,6 +18,8 @@
   },
   "devDependencies": {
     "@firebase/testing": "^0.11.0",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
     "babel-eslint": "^8.2.1",
     "babel-preset-env": "^1.7.0",
     "babel-register": "^6.26.0",
@@ -24,16 +28,21 @@
     "dirty-chai": "^2.0.1",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^4.16.0",
+    "eslint": "^7.25.0",
     "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-compat": "^3.9.0",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-mocha": "^5.2.1",
+    "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-react": "^7.5.1",
     "firebase-bolt": "^0.8.4",
     "mocha": "^5.2.0",
+    "prettier": "^2.2.1",
     "sinon": "^7.3.2",
+    "typescript": "^4.2.4",
     "webpack-cli": "^3.3.4",
     "webpack-dev-server": "^3.7.1"
   },
